+++ util/status_message.cc
@@ -8,10 +8,11 @@
 namespace rocksdb {
 
 const char* Status::msgs[] = {
-    "",                                                  // kNone
-    "Timeout Acquiring Mutex",                           // kMutexTimeout
-    "Timeout waiting to lock key",                       // kLockTimeout
-    "Failed to acquire lock due to max_num_locks limit"  // kLockLimit
+    "",                                                   // kNone
+    "Timeout Acquiring Mutex",                            // kMutexTimeout
+    "Timeout waiting to lock key",                        // kLockTimeout
+    "Failed to acquire lock due to max_num_locks limit",  // kLockLimit
+    "No space left on device"                             // kNoSpace
 };
 
 }  // namespace rocksdb