+++ utilities/column_aware_encoding_exp.cc
@@ -80,8 +80,7 @@ class ColumnAwareEncodingExp {
     std::unordered_map<std::string, CompressionType> compressions = {
         {"kNoCompression", CompressionType::kNoCompression},
         {"kZlibCompression", CompressionType::kZlibCompression},
-        {"kZSTDNotFinalCompression",
-         CompressionType::kZSTDNotFinalCompression}};
+        {"kZSTD", CompressionType::kZSTD}};
 
     // Find Compression
     CompressionType compression_type = compressions[FLAGS_compression_type];